Add API import support through *Imports objects and traits

+++ src/main/scala/sutil/modeling/Counted.scala
@@ -1,40 +1,40 @@
 package sutil.modeling
 
-import org.slf4j.LoggerFactory
 import grizzled.slf4j.Logging
 import scala.collection.mutable.ConcurrentMap
-import java.util.concurrent.atomic.AtomicInteger
 import java.util.concurrent.ConcurrentHashMap
 import scala.collection.JavaConversions._
+import java.util.concurrent.atomic.AtomicLong
 
 object Counted {
 
-  private val counters: ConcurrentMap[Class[_], AtomicInteger] =
-    new ConcurrentHashMap[Class[_], AtomicInteger]
+  private val counters: ConcurrentMap[Class[_], AtomicLong] =
+    new ConcurrentHashMap[Class[_], AtomicLong]
 
-  protected val defaultLoggingMessage: Int ⇒ String =
+  protected val defaultLoggingMessage: Long ⇒ String =
     "%s instances created" format _
 
-  protected val defaultLoggingPeriod: Option[Int] = Some(1000)
+  protected val defaultLoggingPeriod: Option[Long] = Some(1000)
 
 }
 
 trait Counted extends Logging {
 
-  protected val countedLoggingPeriod: Option[Int] =
+  protected val countedLoggingPeriod: Option[Long] =
     Counted.defaultLoggingPeriod
 
-  protected val countedLoggingMessage: Int ⇒ String =
+  protected val countedLoggingMessage: Long ⇒ String =
     Counted.defaultLoggingMessage
 
   protected val countedInstanceNumber =
     Counted.counters
-      .getOrElseUpdate(getClass, new AtomicInteger(0))
-      .incrementAndGet
+    .getOrElseUpdate(getClass, new AtomicLong(0))
+    .incrementAndGet
 
-  countedLoggingPeriod foreach { period ⇒
-    if (countedInstanceNumber % period == 0)
-      info(countedLoggingMessage(countedInstanceNumber))
+  countedLoggingPeriod foreach {
+    period ⇒
+      if (countedInstanceNumber % period == 0)
+        info(countedLoggingMessage(countedInstanceNumber))
   }
 
 }
